requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
kiwisolver==1.4.9
matplotlib==3.10.8
numpy>=1.26.0,<2.3.0
packaging==25.0
pathlib==1.0.1
pillow==12.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
six==1.17.0
torch==2.9.1
torchvision==0.24.1
torchaudio==2.9.1
scikit-learn==1.8.0
opencv-python==4.12.0.88